Make sure current_wordsize and PERSONALITY0_WORDSIZE are ints in all arches
On 64bit systems with a single personality, they used to be sizeof(long),
which has type "long", not "int", which complicates printf formats.

* defs.h: Ensure that PERSONALITY0_WORDSIZE;s tyoe is int.
This in turn makes sure current_wordsize is also an int.
* count.c (call_summary): Revert the change which added cast to int.

+++ count.c
@@ -220,7 +220,7 @@ call_summary(FILE *outf)
 		if (i)
 			fprintf(outf,
 				"System call usage summary for %d bit mode:\n",
-				(int) (current_wordsize * 8));
+				current_wordsize * 8);
 		call_summary_pers(outf);
 	}
